More flake8 plugins
- flake8-expression-complexity
- flake8-requirements

Adding the requirements plugin required specifying a couple of modules
that it was never going to be able to pick up. "rediscluster" comes from
the "redis-py-cluster" package (thanks), and I had to allow the "tests"
package through too.

I also experimented with flake8-annotations-coverage, but ultimately
decided to leave it out for the time being. Why? For similar reasons to
why I haven't turned on mypy's strict mode. It wants an inordinate
number of useless void return type hints on test functions.